Implement facility for on-board animation when no serial input.
Main-loop follows shift-out and serial handlers.

When serial goes idle, it start generating frames, in sync with frame
shift-out.

When serial is enabled again, it switches back to show what is received.

As a bonus, we now auto-re-sync with serial when it has been idle for a bit.

(skeleton on-board animation only for now)