Changes to work with Arduino Uno.
Seems Uno can loose a few bytes on USB-to-serial at reset. So add a sleep,
and check for error reports - and if error, sleep a bit to get back in
sync.

Also work-around strange delay-0.2sec-every-10-sec on non-blocking reading
of the serial port, by spawning a child to proxy the reads onto a pipe.

Uno seems to drive the cube stable now.