Fix sync-up algorithm on the LED-cube.
The old method was basically just kind-of-broken.

Now we check the time passed since last byte received on the serial line.
When 60 refresh frames (around 1/3 seconds) have passed without any
serial activity we reset the state - this way we're in sync with the start of
a frame when data starts arriving again.